sw: online: make sure commment removal notification is sent
problem:
when comment is added inside a table in online,
and entire table is deleted, comments were not removed
due to notification not being sent.

Another failed case was when some text with comment is
copy pasted, and then pasted text is undoed,
text was removed but comment stayed.

This patch will notify comment removal before actaully distroying the comment/note

Signed-off-by: Pranam Lashkari <lpranam@collabora.com>
Change-Id: I70aad7eb421b4bde52b9e54855e2360c59807852
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/138398
Reviewed-by: Szymon Kłos <szymon.klos@collabora.com>
Tested-by: Jenkins CollaboraOffice <jenkinscollaboraoffice@gmail.com>
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/138521
Tested-by: Jenkins