tdf#144092 pptx import: correct table height during import
It appears PowerPoint can export rows of a table with row heights that
is less than the minimum height (for that row). And also export the
total table height wrong with it.

If PowerPoint imports such a table, those rows are individually
expanded to the minimum height. (Also increasing the table's total
height)

In Impress import we calculate table height by adding up individual
row heights. Table layouting code depends on the table height being
correct. This is why rows with less than minimum height lead to
layouting problems.

To compensate for this, while importing tables, layouting is skipped
until the table height is updated with the corrected height.

The correct height is calculated by layouting the table without
fitting to an area (i.e with bFit = false).

Change-Id: I79187882470a4e285b45bca1eabb469a084067f5
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/138652
Tested-by: Jenkins
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>