tdf#140215 Simplify the handling of .ulf files
Now the headings in the ulf files for .desktop files are
in the form [filename_Key]

Gallery names are also adjusted to fit the new scheme, where there is
no longer a need to pass a --key argument to desktop-translate.py

Sync comments with .desktop files and the remaining .ulf and remove
obsolete Mandriva Linux meta data while at it.

Script to mass-replace relevant names in translations will be
provided to infra.

Change-Id: I87e8028aa5b66f5f5560efa62ddd9b1e5b61c49c
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/138455
Tested-by: Jenkins
Reviewed-by: Sophie Gautier <sophi@libreoffice.org>
Reviewed-by: Christian Lohmaier <lohmaier+LibreOffice@googlemail.com>