chart2: always use the series for the keys (symbols) in data table
In special case the legend keys are not series, but can be each
value of a series. This can't be done for the data table where so
this has been disabled.

This caused a crash when positioning the symbols, as in some cases
we got 0 symbols (due to another bug that didn't put the symbols
into the collection) or more symbols that there were series (and
the symbols couldn't be positioned correctly). With this change
we now get the correct number of symbols for the data table.

Also added the test document and test to open the document, to
make sure we don't crash anymore.

Change-Id: Idbebdd5295595a823b5420958e803b13258df035
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/138697
Tested-by: Jenkins
Reviewed-by: Tomaž Vajngerl <quikee@gmail.com>