vcl: Fix GenericSalLayout::GetTextWidth()
There was a mismatch between text width calculated in
GenericSalLayout::GetTextWidth() and last glyph width calculated in
GenericSalLayout::GetCharWidths(), with the later using the sum of
GlyphItem::newWidth() and the later using the GlyphItem::linearPos(),
but linearPos() includes glyph offset and it should be subtracted from
it to get actual width and we were subtracting it in the wrong place.

I fixed where the xOffset() is subtracted but didn’t change it to simply
sum GlyphItem::newWidth() though that is simpler so that we continue to
test that linearPos() is correctly set.

Change-Id: I40d6be6d1c82ed0ca1ce04bf792adf72faa0d787
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/139071
Tested-by: Jenkins
Reviewed-by: خالد حسني <khaled@aliftype.com>