crashtesting: fix DOCX export of forum-mso-de-102317.docx
This went wrong in b2bc4ea8ddae6b01f344469d988e21fed3143c42 (DOCX
import: handle SDT around citation field, 2014-07-16), the problem is
that a run/inline SDT is pending, but we wrote a header stream in the
meantime, so the invariant that Start/EndParagraph() in
DocxAttributeOutput is called in pairs doesn't help, which results in
not-well-formed XML, which is an assertion failure in debug build.

Fix the problem similar to 5f3af56b2c0ef6c628a7cfe5ce6e86f8e1765f5f (sw:
refactoring for docx sdt blocks output, 2021-11-16), which already did
the same for paragraph/block SDTs.

Change-Id: I5e7273b839bb5c1f841c0fab9b68b4ca239cca18
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/139089
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Tested-by: Jenkins