sw HTML import: fix height of OLE objs when it is missing and width is relative
In case an OLE object's fallback bitmap had a width=100% size but no
explicit height, we used to set the OLE object's size to a fixed value,
based on the bitmap's pixel size.

Falling back to the pixel size based on the current DPI is sensible if
there is no height specified, but browsers keep the aspect ratio when
only width or height is specified and we didn't.

Fix the problem by explicitly checking for the "one axis has percent
size, the other is unset" scenario in SwHTMLParser::InsertEmbed() and
then handling the new SwFormatFrameSize::SYNCED value in
SwHTMLParser::SetFixSize().

Note that SetFixSize() is not used for bitmaps (only for OLE objects),
so that codepath may still need a similar fix.

Change-Id: I9f70307edbfbb06e228494fb5c04e7381ed8e5cc
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/139526
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Tested-by: Jenkins