tdf#149990 Remove unwarranted, unlocalizable term from main launcher
This change was added in commit 6ed597a7dd1da40248236e7a71843c2d76d6173e
for unclear reasons (that wasn’t asked for in the bug report), and,
besides, it solves an inexistent problem. No one has been confused on their
expectations after clicking this launcher on their app selector or Start menu.

For anyone looking, it is an obvious thing that, if presented with a set of
launchers like:

   LibreOffice
   LibreOffice Writer
   LibreOffice Calc
   ...

... the option that isn’t specific will launch a general interface. That’s just simple.

“Start Center” is not the brand name of an application. Although, technically, it is
a module, for users, it’s just a welcome screen, to which OOo has added a glorified
name (it seems we’re experts on inventing weird terms for everyday UI paradigms…
just see “NotebookBar”, which is just a ribbon; deal with it).

It shouldn’t be present in the main launcher for the same reason we don’t add
the word “Dialog” to the title of every dialog.

Change-Id: I3703299fcb8a8244c721631c3107c9f9d3264d83
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/139805
Tested-by: Jenkins
Reviewed-by: Adolfo Jayme Barrientos <fitojb@ubuntu.com>