tdf#103829 ods export: don't output invalid office:string-value
If the formatted (i.e. \t) string was different from
the raw string (i.e. \x001 EE_FEATURE_TAB marker),
then export was outputing office:string-value
with IsInvalidChar \x001.

The impact on import is that the cell wants to display
this string-value, which is missing all of the "invalid char"
stuff. So, just don't export a wrong string at all.

Change-Id: Ib98e58d0c6e1186d199a4dcf294b56403b710f48
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/139816
Tested-by: Jenkins
Reviewed-by: Justin Luth <jluth@mail.com>