Related: tdf#149665 Unify input of a leading ' apostrophe in non-Text cell
If a cell is not formatted as Text and
ScSetStringParam::mbHandleApostrophe is true, all input (except
one single apostrophe that otherwise would lead to an empty cell
string) now is treated the same and the leading apostrophe is
removed and the remainder is set as text content. The Input Line
and editing a cell get an apostrophe prepended for the cases
needed to generate such escaped string again.

This made it necessary to adapt a unit test that exactly checked
the old behaviour.

Change-Id: I07b35d33f3704970784dc6de1322bda28903bb97
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/139934
Reviewed-by: Eike Rathke <erack@redhat.com>
Tested-by: Jenkins