tdf#75702 ODS import: send text:line-break to editeng
A cell with a line-break specified is no longer a simple string,
so send the string to editeng where it can be properly displayed.

This already works fine for XLS and XLSX.

This change effectively means that text:line-break
will disappear, and become two separate paragraphs.
Given the nature of spreadsheets, the precidence in
various other patches and the failure to properly
display without this, I think I'm fine with
unilaterally changing one paragraph into two.

Change-Id: I7abcbc6f698a4cb319aea5465a8b0393cf63f31f
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/140262
Tested-by: Jenkins
Reviewed-by: Justin Luth <jluth@mail.com>